gypsum: Introduce ModuleScope for per-file bindings
ModuleScope is a scope for each file in a package. Global definitions
and imports are bound in ModuleScope initially. Definitions (but not
imports) are copied to GlobalScope after we have walked the AST.

This means that symbols that are imported in one file are not visible
in other files. Imported symbols shadow definitions in other files but
they are overloaded with definitions in the same file.

Fixes #34.